Add conditional check before remove container

+++ destroyNginx.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 NGINX_NAME=${NGINX_NAME:-proxy}
-docker stop ${NGINX_NAME}
-docker rm -v ${NGINX_NAME}
+if [ -n "$(docker ps -a | grep ${NGINX_NAME})" ]; then
+  docker stop ${NGINX_NAME}
+  docker rm -v ${NGINX_NAME}
+fi
